requirements.txt
# Core
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
scipy>=1.5.0
scikit-image>=0.17.0
tqdm>=4.50.0
Pillow>=8.0.0

# Medical Imaging
nibabel>=3.2.0
SimpleITK>=2.0.0
pydicom>=2.1.0
dicom2nifti>=2.2.0

# Image Registration (optional)
antspyx>=0.2.0

# Deep Learning Utilities
wandb>=0.12.0
tensorboard>=2.4.0
lpips>=0.1.4
pytorch-fid>=0.2.0
kornia>=0.5.0

# Image Processing
opencv-python>=4.5.0
albumentations>=1.0.0

# Hyperparameter Optimization
optuna>=2.8.0

# Utilities
pyyaml>=5.4.0
jsonschema>=3.2.0